chore: rename js modules to ui
Addresses #49
