+++ themes/thecodeboss/assets/js/ui/typing.ts
